Move session stuff to common
