+++ e2e-tests/src/staking.rs
@@ -1,5 +1,4 @@
-use crate::session::wait_for_session;
-use common::{create_connection, send_xt, BlockNumber, Connection, KeyPair};
+use common::{wait_for_session, create_connection, send_xt, BlockNumber, Connection, KeyPair};
 use log::info;
 use pallet_staking::{RewardDestination, ValidatorPrefs};
 use primitives::Balance;
